* Refactoring * Test fixing

+++ src/test/java/de/tubs/cs/ias/asm_test/taintaware/range/testHelper/RangeChainer.java
@@ -37,7 +37,7 @@ public class RangeChainer {
             throw new IndexOutOfBoundsException(Integer.toString(source));
         }
 
-        ranges.add(new IASTaintRange(start, end, (short) source));
+        ranges.add(new IASTaintRange(start, end, source));
 
         return this;
     }
